Set visibilityString from constructor
