+++ src/main/java/teammates/ui/template/FeedbackResponseRow.java
@@ -55,8 +55,7 @@ public class FeedbackResponseRow {
                 Map<String, String> instructorEmailNameTable = results.instructorEmailNameTable;
                 FeedbackResponseCommentRow responseRow = new FeedbackResponseCommentRow(frc,
                         giverEmail, giverName, recipientName, showCommentTo, showGiverNameToString, responseVisibilities,
-                        instructorEmailNameTable, results.getTimeZone());
-                responseRow.setVisibilityIconString(getTypeOfPeopleCanViewComment(frc, question));
+                        instructorEmailNameTable, results.getTimeZone(), question);
                 responseRow.enableEditDelete();
                 this.responseComments.add(responseRow);
             }
@@ -81,59 +80,6 @@ public class FeedbackResponseRow {
 
     public List<FeedbackResponseCommentRow> getResponseComments() {
         return responseComments;
-    }
-
-    /**
-     * Returns the type of people that can view the response comment.
-     */
-    public String getTypeOfPeopleCanViewComment(FeedbackResponseCommentAttributes comment,
-                                                FeedbackQuestionAttributes relatedQuestion) {
-        List<FeedbackParticipantType> showCommentTo;
-        if (comment.isVisibilityFollowingFeedbackQuestion) {
-            showCommentTo = relatedQuestion.showResponsesTo;
-        } else {
-            showCommentTo = comment.showCommentTo;
-        }
-        if (showCommentTo == null || showCommentTo.isEmpty()) {
-            return "nobody";
-        }
-
-        StringBuilder peopleCanView = new StringBuilder(100);
-        for (int i = 0; i < showCommentTo.size(); i++) {
-            FeedbackParticipantType commentViewer = showCommentTo.get(i);
-            if (i == showCommentTo.size() - 1 && showCommentTo.size() > 1) {
-                peopleCanView.append("and ");
-            }
-
-            switch (commentViewer) {
-            case GIVER:
-                peopleCanView.append("response giver, ");
-                break;
-            case RECEIVER:
-                peopleCanView.append("response recipient, ");
-                break;
-            case OWN_TEAM:
-                peopleCanView.append("response giver's team, ");
-                break;
-            case RECEIVER_TEAM_MEMBERS:
-                peopleCanView.append("response recipient's team, ");
-                break;
-            case STUDENTS:
-                peopleCanView.append("other students in this course, ");
-                break;
-            case INSTRUCTORS:
-                peopleCanView.append("instructors, ");
-                break;
-            default:
-                break;
-            }
-        }
-        String peopleCanViewString = peopleCanView.toString();
-        return removeEndComma(peopleCanViewString);
-    }
-
-    private String removeEndComma(String str) {
-        return str.substring(0, str.length() - 2);
     }
 
 }
